fix(folder structure): interface file moved to domain folder
